Add `u64`-based and native implementations of polyfills
