+++ crates/farmer/ab-proof-of-space-gpu/src/shader/num/native/tests.rs
@@ -1,0 +1,1 @@
+include!("../u32/tests.rs");
